Uint to Int

+++ android/src/main/kotlin/dev/steenbakker/nordicdfu/NordicDfuPlugin.kt
@@ -71,7 +71,7 @@ class NordicDfuPlugin : FlutterPlugin, MethodCallHandler, EventChannel.StreamHan
         val address = call.argument<String>("address")
         val name = call.argument<String>("name")
         var filePath = call.argument<String>("filePath")
-        val maxMtu = call.argument<UInt>("maxMtu")
+        val maxMtu = call.argument<Int>("maxMtu")
         var fileInAsset = call.argument<Boolean>("fileInAsset")
         val forceDfu = call.argument<Boolean>("forceDfu")
         val enableUnsafeExperimentalButtonlessServiceInSecureDfu =
@@ -139,7 +139,7 @@ class NordicDfuPlugin : FlutterPlugin, MethodCallHandler, EventChannel.StreamHan
         address: String,
         name: String?,
         filePath: String,
-        maxMtu: UInt?,
+        maxMtu: Int?,
         forceDfu: Boolean?,
         enableUnsafeExperimentalButtonlessServiceInSecureDfu: Boolean?,
         disableNotification: Boolean?,
@@ -163,7 +163,7 @@ class NordicDfuPlugin : FlutterPlugin, MethodCallHandler, EventChannel.StreamHan
             )
         }
         if(maxMtu != null){
-            starter.setMtu(maxMtu.toInt())
+            starter.setMtu(maxMtu)
         }
         if (forceDfu != null) starter.setForceDfu(forceDfu)
         if (disableNotification != null) starter.setDisableNotification(disableNotification)
